Remove ordering requirement for hashCode/equal. Bump version to 1.0.3-SNAPSHOT.
Now if #hashCode() is defined before #equals() in the class, the #equals() method
is found and type checked, forcing a macro expansion, and then the attachments are
taken from there. Slightly inefficient because it needs to expand the macro twice,
but probably not a big deal.

Additionally, modified some tests to prove ordering no longer affects macro expansion